✨ Implement matching icons for each item
This workflow presents a list of emoji, so naturally the icon for each
list item would be that emoji. Problem is, Alfred only supports files
for this.

What we can do is grab the pngs we need from EmojiOne. I'm having them
written to the data directory, which seems better than just packaging
them in the workflow?